+++ spec/eo_time_spec.rb
@@ -537,6 +537,27 @@ describe EtOrbi::EoTime do
     end
   end
 
+  describe '#rweek, #rday' do
+
+    {
+
+      '2008-12-31 12:00 Europe/Lisbon' => [ 3, 366, -521, -3652 ],
+      '2018-12-31 12:00 Europe/London' => [ 1, 365, 0, 0 ],
+      '2019-01-01 12:00 Europe/Paris' => [ 2, 1, 1, 1 ],
+      '2019-04-02 12:00 Europe/Berlin' => [ 2, 92, 14, 92 ],
+
+    }.each do |t, (wday, yday, rweek, rday)|
+
+      it "for #{t}, returns rweek:#{rweek}/rday:#{rday}" do
+
+        t = EtOrbi.make_time(t)
+
+        expect([ t.wday, t.yday, t.rweek, t.rday ])
+          .to eq([ wday, yday, rweek, rday ])
+      end
+    end
+  end
+
   describe '#localtime' do
 
     it 'returns a new EoTime instance in the local zone' do